flatten FileTime service facade into time.ts
